requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pyserial>=3.5

